% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_play.R
\name{scatter_play}
\alias{scatter_play}
\title{ScatterPlay}
\usage{
scatter_play()
}
\description{
Manually scatter data points to see how correlation and regression are affected
}
\examples{
scatter_play()
}

